Day10
The program must accept an integer N as the input. The program must print the sum of permutations of the digits in N as the output.

Boundary Condition(s):
10 <= N <= 10^8

Input Format: The first line contains N.
Output Format: The first line contains the sum of permutations of the digits in N.

Example Input/Output 1:
Input: 
231

Output: 
1332

Explanation:
The permutations of the digits in 231 are 231, 213, 123, 132, 321 and 312. So the sum is 1332 (231+213+123+132+321+312). Hence the output is 1332

Example Input/Output 2:
Input:
7865

Output:
173316