Day15
The program must accept N integers as the input. The program must sort the integers based on thenumber of ones in their binary representation in descending order. If more than one integers havethe same number of ones in their binary representation, the program must sort those integers inascending order. Finally, the program must print the N sorted integers as the output.

Boundary Condition(s):
2 <= N <= 1000
1 <= Each integer value <= 10^8

Input Format:
The first line contains N.
The second line contains N integers separated by a space.

Output Format:
The first line contains the N sorted integers separated by a space.

Example Input/Output 1:
Input:
6
65 2 78 790 10 31

Output:
31 790 78 10 65 2

Explanation:
The binary representation of 31 is 11111.
The binary representation of 790 is 1100010110.
The binary representation of 78 is 1001110.
The binary representation of 10 is 1010.
The binary representation of 65 is 1000001.
The binary representation of 2 is 10.
Here the 6 integers are sorted based on the number of ones in their binary representation.

Example Input/Output 2:
Input:
5
9 15 6 4 14

Output:
15 14 6 9 4