Top, right, left, bottom adjacent cells
The program must accept an integer matrix of size NxN as the input. For eachoccurrence of 0 in the matrix, the program must find the sum S of integers in thetop, right, bottom and left adjacent cells. Then the program must replace 0 with thesum S and the four adjacent cells with 0. Finally, the program must print the modified matrix as the output.Note: The zeros are placed in a matrix so that the four adjacent cells of zero cannot be adjacent to the another zero.

Boundary Condition(s):
3 <= N <= 50

Input Format:
The first line contains N.
The next N lines, each containing N integers separated by a space.

Output Format:
The first N lines containing the modified matrix.

Example Input/Output 1:
Input:
5

2 8 3 9 2

6 9 2 3 9

2 3 2 0 1

1 0 4 6 3

5 7 4 7 6

Output:
2  8 3  9 2

6  9 2  0 9

2  0 0 12 0

0 15 0  0 3

5  0 4  7 6