Extracting Month
The program must accept a valid date D in the format of DD-MM-YYYY as the input. Theprogram must print the name of the month in the given date as the output.

Boundary Condition(s):
1 <= DD <= 31
1 <= MM <= 12
1000 <= YYYY <= 2020

Input Format:
The first line contains a valid date.

Output Format:
The first line contains the month for the given date.

Example Input/Output 1:
Input:
27-11-1997

Output:
November

Explanation:
The given date is 27-11-1997. The month 11 indicates November.
Hence the output is November

Example Input/Output 2:
Input:

08-04-2020

Output:
April