Step Numbers from 1 to N
The program must accept an integer N as the input. The program must print all the step numbersfrom 1 to N as the output. If there is no such integer then the program must print -1 as the output.A number is called a step number if all the adjacent digits have an absolute difference of 1.Note: The number of digits in a step number is always greater than 1.

Boundary Condition(s):
1 <= N <= 10^7

Input Format:
The first line contains N.

Output Format:
The first line contains the integer value(s) or -1.

Example Input/Output 1:
Input:
12

Output:
10 12

Explanation:
The absolute difference between all the adjacent digits in the integer 10 is 1 (1 - 0).The absolute difference between all the adjacent digits in the integer 12 is 1 (1 - 2).
So the step numbers from 1 to 12 are 10 and 12.
Hence the output is 10 12

Example Input/Output 2:
Input:
5

Output:
-1

Example Input/Output 3:
Input:
130

Output
10 12 21 23 32 34 43 45 54 56 65 67 76 78 87 89 98 101 121 123