Next Greater - Equal 1's Binary
The program must accept an integer N as the input. The program must find the numberof 1's in the binary representation of N. Then the program must print the next greaterinteger of N having the same number of 1's in the binary representation.

Boundary Condition(s):
1 <= N <= 10^5

Input Format:
The first line contains N.

Output Format:
The first line contains an integer.

Example Input/Output 1:
Input:
6

Output:
9

Explanation:
The binary representation of 6 is 110. The count 1's in 110 is 2.
So the next greater integer with two 1's in the binary representation is 9. The binary representation of 9 is 1001.

Example Input/Output 2:
Input:
13

Output:
14

Example Input/Output 3:
Input:
15

Output:
23