Palindrome - Decimal & Binary
The program must accept an integer N as the input. The program must print yes if N and its binary representation are palindrome. Else the program must print no as the output.

Boundary Condition(s):
1 <= N <= 10^8

Input Format:
The first line contains N.

Output Format:
The first line contains yes or no.

Example Input/Output 1:
Input: 
585

Output: 
yes

Explanation:
Here N = 585.
The integer 585 is a palindrome.
The binary representation of the integer 585 is 1001001001, which is also a palindrome. So yes is printed as the output.

Example Input/Output 2:
Input:
2772

Output:
no