Patients Token
The program must accept the names of N patients and the time (in 24-hr format) they entered a hospital as theinput. The hospital's policy is to provide tokens based on the patient's entry time. The program must print thenames of the N patients along with their token number in ascending order.
Note: Two entries are not allowed at a given time.

Boundary Condition(s):
3 <= N <= 50
3 <= Length of each patient's name <= 50

Input Format:
The first line contains N.
The next N lines, each containing the name and the entry time of a patient separated by a space.

Output Format:
The first N lines, each containing the name and the token number of a patient separated by a space.

Example Input/Output 1:
Input:
5

Sandy 09:30
Joel 11:15
Priya 21:00
Abdul 13:46
Kapli 13:45

Output:

Sandy 1
Joel 2
Kapli 3
Abdul 4
Priya 5

Explanation:
The tokens are provided based on the entry time of the patients.
The patient Sandy entered the hospital at 09:30, so he has the token 1.
The patient Joel entered the hospital at 11:15, so he has the token 2.The patient Kapli entered the hospital at 13:45, so he has the token 3.
The patient Abdul entered the hospital at 13:46, so he has the token 4.The patient Priya entered the hospital at 21:00, so she has the token 5.

Hence the output is
Sandy 1
Joel 2
Kapli 3
Abdul 4
Priya 5