Prime Integer - Add Smallest Value
The program must accept N integers as the input. For each integer value X among the N integers,the program must print the smallest value that can be added to the integer X (possibly zero) so thatthe integer becomes a prime integer.

Note: Optimize your logic to avoid Time Limit Exceeded Error.

Boundary Condition(s):
1 <= N <= 2,00,000
2 <= Each integer value <= 10^6

Input Format:
The first line contains N.
The second line contains N integers separated by a space.

Output Format:
The first line contains N integers separated by a space.

Example Input/Output 1:
Input:
7
89 54 36 74 44 19 12

Output:
0 5 1 5 3 0 1

Explanation:
The integer 89 is a prime integer. So the smallest value to be added is 0.
The smallest value to be added to 54 is 5. So the integer becomes 59, which is a prime integer.
The smallest value to be added to 36 is 1. So the integer becomes 37, which is a prime integer.The smallest value to be added to 74 is 5. So the integer becomes 79, which is a prime integer.
The smallest value to be added to 44 is 3. So the integer becomes 47, which is a prime integer.The integer 19 is a prime integer. So the smallest value to be added is 0.
The smallest value to be added to 12 is 1. So the integer becomes 13, which is a prime integer.
Hence the output is 0 5 1 5 3 0 1

Example Input/Output 2:
Input:
9
26 10 88 49 20 71 25 83 3

Output:
3 1 1 4 3 0 4 0 0