Spiral Diagonal Pattern
The program must accept a string S as the input. The program must generate a matrix with asterisks and replace theasterisks in the diagonals of the matrix with the characters in the string S spirally (in the clockwise direction). Then theprogram must print the modified matrix as shown in the Example Input/Output section.

Boundary Condition(s):
5 <= Length of S <= 100

Input Format:
The first line contains S.

Output Format:
The lines containing the modified matrix.

Example Input/Output 1:
Input:
skillrack

Output:

S***k
*l*r*
**k**
*c*a*
l***i

Explanation:
After generating the matrix with asterisks, the diagonals are replaced with the characters in the string skillrack.The first 4 characters in the string skillrack are s, k, i, I. They are placed in the outermost layer in the clockwise direction along the diagonals.
Then the next 4 characters in the string skillrack are I, r, a, c. They are placed in the next inner layer.
Then the remaining character k is placed in the innermost layer.
S***k
*l*r*
**k**
*c*a*
l***i

Example Input/Output 2:
Input:
Computer

Output:
C**o
*ut*
*re*
p**m