Alphabets and Sum of Digits
The program must accept a string S containing only alphabets and digits as the input. Theprogram must print the alphabets and then the sum of the digits in the string S as the output.Note: At least one alphabet and one digit are always present in the string S.

Boundary Condition(s):
2 <= Length of S <= 1000

Input Format:
The first line contains S.

Output Format:
The first line contains the alphabets in S followed by the sum of the digits in S.

Example Input/Output 1:
Input:
A1B2c3d4E5

Output:
ABcdE15

Explanation:
The alphabets in the string A1B2c3d4E5 are A, B, c, d and E. So they are printed.
The digits in the string A1B2c3d4E5 are 1, 2, 3, 4 and 5. So their sum 15 is printed.