Missing digit - Divisible by 8
A positive integer N that has one of its digit missing is passed as the input. The missing digit is marked as X. The program must print the smallest possible digit that can replace X so that N is divisible by 8 as the output. If such a digit is not possible, the program must print -1 as the output.
Note: There are no leading zeroes in the integer N.

Boundary Condition(s):
100 <= N <= 10^8

Input Format:
The first line contains N.

Output Format:
The first line contains the missing digit or -1.

Example Input/Output 1:
Input:
543X0

Output:
2

Explanation:
Here N = 543X0.
If X is replaced with 2 or 6, the integer N is divisible by 8. For any other value of X, the integer N is not divisible by 8.
The smallest possible digit that can replace X is 2. So 2 is printed as the output.

Example Input/Output 2:
Input:
12X515

Output:
-1

Example Input/Output 3:
Input:
X16

Output:
2