Boxes - Alternate Left and Right
There are N boxes arranged circularly, where each box contains gold rings. The number of gold rings in each boxis passed as the input to the program. If the number of gold rings in a box is odd, K gold rings are removed fromthat box and they are distributed to it's adjacent boxes alternatively (left and right). When K gold rings are notpresent in a box, it must be kept as it is. The value of K is also passed as the input to the program. The programmust perform the above process on the given N boxes and print the number of gold rings in the N boxes as theoutput.

Boundary Condition(s):
3 <= N <= 100
1 <= Each integer value, K <= 10^5

Input Format:
The first line contains N.
The second line contains N integers separated by a space.
The third line contains K.

Output Format:
The first line contains N integers separated by a space.

Example Input/Output 1:
Input:
5
5 3 6 1 7
3

Output:
3 4 6 3 6

Explanation
Here N=5 and K=3, the number of gold rings in 5 boxes are 5, 3, 6, 1 and 7.The number of gold rings in the first box is 5, which is odd. So 3 gold rings are removed from that box and added 2 gold rings to it's adjacent left box (5th box) and added 1 gold ring to it's adjacent right box (2nd box).
Afterdistributing the gold rings, the number of gold rings in the 5 boxes are 2, 4, 6, 1 and 9.
The number of gold rings in the second box is 4, which is even. So it must be kept as it is.The number of gold rings in the third box is 6, which is even. So it must be kept as it is.
The number of gold rings in the fourth box is 1, which is less than 3. So it must be kept as it is.The number of gold rings in the fifth box is 9, which is odd. So 3 gold rings are removed from that box and added 2 gold rings to it's adjacent left box (4th box) and added 1 gold ring to it's adjacent right box (1st box).After distributing the gold rings, the number of gold rings in the 5 boxes are 3, 4, 6, 3 and 6.
Hence the output is 3 4 6 3 6

Example Input/Output 2:
Input:
8
17 5 9 12 8 20 14 21
6

Output:
14 11 6 12 8 20 14 21