Bishop Movement
There is an RxC chessboard. The chessboard contains one bishop, one pawn and empty squares. The indices of the bishop (X, Y) and the pawn (M, N) are passed as the input to the program. The bishop moves in any direction diagonally. If the pawn is encountered or meets the end of the board, the movement of bishop stops. The program must print the number of squares that the bishop can move as the output. 

Boundary Condition(s): 
2 <= R, C <= 100 0 <= X, M < R 0 <= Y, N < C 

Input Format: 
The first line contains R and C separated by a space. 
The second line contains X and Y separated by a space. 
The third line contains M and N separated by a space. 

Output Format: 
The first line contains the number of squares that the bishop can move.

Example Input/Output 1: 
Input: 
7 7 
3 3 
6 0 

Output: 
11 

Explanation: 
The bishop movable squares are highlighted below. 

* * * * * * * 
* * * * * * * 
* * * * * * * 
* * * B * * * 
* * * * * * * 
* * * * * * * 
P * * * * * * 

The asterisks represent the empty squares. 
B represents the bishop. P represents the pawn. 
The number of squares that bishop can move is 11. 
So it is printed as the output. 

Example Input/Output 2: 
Input: 
5 6 
1 1 
3 1 

Output: 
6