Replace Max Digit - Two Integers
The program must accept two integers X and Y as the input. The program must replace all the occurrences of the maximum digit in X with the maximum digit in Y. Then the program must print the sum of X and Y as the output. 

Boundary Condition(s): 
10 <= X, Y <= 10^8 

Input Format: 
The first line contains X and Y separated by a space. 

Output Format: 
The first line contains an integer representing the sum of X and Y. 

Example Input/Output 1: Input: 
1878 2020 

Output: 3292 

Explanation: 
Here X = 1878 and Y = 2020. 
After replacing all the occurrences of the maximum digit 8 in 1878 with the maximum digit 2 in 2020, the integer 1878 becomes 1272. 
Here the sum of 1272 and 2020 is 3292. 
Hence the output is 3292 

Example Input/Output 2: Input: 
1111 144 

Output: 
4588 

Example Input/Output 3: Input: 
189 1239 

Output: 
1428