Twice and Reverse
The program must accept two integers N and T as the input. The program must print the integers from N to 1 inthe 1st row and then in the 2nd row the values must be twice that of the values in the 15t row in reverse order.Then in the 3rd row the values must be twice that of the values in the 2nd row. Similarly, the program must repeatthe process for T rows and print the integer values.

Boundary Condition(s):
2 <= N, T <= 30

Input Format:
The first line contains N and T separated by a space.

Output Format:
The first T lines, each contains N integers separated by a space.

Example Input/Output 1:
Input:
4 5

Output:
4 3 2 1
2 4 6 8
16 12 8 4
8 16 24 32
64 48 32 16

Explanation:
Here N = 4 and T = 5.
The 1st row contains 4 integers 4 3 2 1 from 4 to 1.
The 2nd row contains 4 integers 2 4 6 8, which are twice the values in the 1st row in the reverse order.
The 3rd row contains 4 integers 16 12 8 4, which are twice the values in the 2nd row in the reverse order.
The 4th row contains 4 integers 8 16 24 32, which are twice the values in the 3rd row in the reverse order.The 5th row contains 4 integers 64 48 32 16, which are twice the values in the 4th row in the reverse order.

Example Input/Output 2:
Input:
5 4

Output:
5 4 3 2 1
2 4 6 8 10
20 16 12 8 4
8 16 24 32 40