Flipping Bits 32-bit Format
The program must accept N positive integers as the input. For each integer X among the given N integers, the program must find the binary representation of X in 32 bits. Then the program must toggle the 32 bits and print its decimal equivalent as the output. 

Boundary Condition(s): 
1 <= N <= 1000 1 <= Each integer value <= 10^9 

Input Format: 
The first line contains N. 
The second line contains N integers separated by a space. 

Output Format: 
The first N lines, each contains an integer. 

Example Input/Output 1: 
Input: 
2 
399861 7 

Output: 
4294567434 
4294967288 

Explanation: 
The binary representation of 399861 in 32 bits is 00000000000001100001100111110101. 
After toggling the bits, it becomes 11111111111110011110011000001010. 
The decimal equivalent of the above binary representation is 4294567434. So it is printed. 
The binary representation of 7 in 32 bits is 00000000000000000000000000000111. 
After toggling the bits, it becomes 11111111111111111111111111111000. 
The decimal equivalent of the above binary representation is 4294967288. So it is printed. 

Example Input/Output 2: 
Input: 
3 
52568330 45 82730988 

Output: 
4242398965 
4294967250 
4212236307 

Max Execution Time Limit: 500 millisecs