Remove Characters - Valid String
The program must accept a string S, a character CH and an integer K as the input. The string is said to be valid only if the character CH is consecutively repeated for at most K times. The program must print the minimum number of characters to be removed from S to become a valid string as the output.

Boundary Condition(s):
1 <= Length of S <= 1000 
1 <= K <= Length of S

Input Format:
The first line contains S, CH and K separated by a space.

Output Format:
The first line contains the minimum number of characters to be removed from S to become a valid string.

Example Input/Output 1:
Input:
abccccddeee c 2

Output:
2

Explanation:
Here 4 consecutive c's have occurred in the string "abccccddeee".
The possible valid string values are
"abddeee" -> 4 c's removed
"abcddeee" -> 3 c's removed
"abccddeee" -> 2 c's removed
So the minimum number of c's to be removed from S is 2.
Hence the output is 2

Example Input/Output 2:
Input: 
BAAABBBBBCCBBBBDBBB B 3

Output:
3