Increment Submatrices
The program must accept the size of an integer matrix RxC and Q queries as the input. Each query contains four integers representing the positions of the top-left and the bottom-right cell of a submatrix. The program must form an integer matrix of size RxC with 0's. For each query, the program must increment the values of the integers in the submatrix specified by the query. Finally, the program must print the modified matrix as the output.
Boundary Condition(s): 
2 <= R, C <= 50
1 <= Q <= 100

Input Format:
The first line contains R and C separated by a space.
The second line contains Q.
The next Q lines, each containing four integers separated by a space.

Output Format:
The first R lines, each containing C integers separated by a space.

Example Input/Output 1:
Input:
4 4
2
1 1 3 2
2 2 3 3

Output:
1 1 0 0
1 2 1 0
1 2 1 0
0 0 0 0

Explanation:
Initially, the value of each cell in the 4x4 matrix is 0.
After incrementing the values in the first submatrix, the matrix becomes
1 1 0 0
1 1 0 0  
1 1 0 0
0 0 0 0
After incrementing the values in the second submatrix, the matrix becomes
1 1 0 0
1 2 1 0
1 2 1 0
0 0 0 0