Replace Vowels with 1
The program must accept a string S as the input. The program must replace all the vowels in thestring S with 1. Then the program must print the modified string and the count of vowels C in thestring S as the output. If the string S contains characters other than alphabets, the program mustprint Invalid as the output.

Boundary Condition(s):
1 <= Length of S <= 1000

Input Format:
The first line contains S.

Output Format:
The first line contains the modified string and C separated by a space or contains the string value "Invalid".

Example Input/Output 1:
Input:
AeFloT

Output:
11FI1T 3

Explanation:
The count of vowels in the string AeFloT is 3 and they are A, e and o.After replacing the vowels with 1, the string becomes 11FI1T.
Hence the output is 11FI1T 3

Example Input/Output 2:
Input:
AB1E

Output:
Invalid